#38 extract inner classes
