+++ app/models/wrapper/AttributeConstraintWrapper.java
@@ -1,0 +1,8 @@
+package models.wrapper;
+
+public enum AttributeConstraintWrapper {
+    NONE,
+    UNIQUE,
+    COMBINATION_UNIQUE,
+    SAME_FOR_ALL;
+}
