+++ app/models/wrapper/ProductDraftWrapper.java
@@ -76,7 +76,7 @@ public class ProductDraftWrapper extends Base {
 
         final List<ImageWrapper> imageWrappers = variantWrapper.getImages();
         final List<Image> images = imageWrappers.stream().map(imageWrapper ->
-                Image.ofWidthAndHeight(imageWrapper.url, imageWrapper.getDimensions().getWidth(),
+                Image.ofWidthAndHeight(imageWrapper.getUrl(), imageWrapper.getDimensions().getWidth(),
                         imageWrapper.getDimensions().getHeight(),
                         "donuts image label")).collect(Collectors.toList());
 
@@ -108,158 +108,5 @@ public class ProductDraftWrapper extends Base {
 
     public List<ProductVariantDraftWrapper> getVariants() {
         return variants;
-    }
-
-    public static class ProductVariantDraftWrapper {
-
-        private final String sku;
-        private final List<PriceWrapper> prices;
-        private final List<AttributeDraftWrapper> attributes;
-        private final List<ImageWrapper> images;
-        private final AvailabilityWrapper availability;
-
-        public ProductVariantDraftWrapper(@JsonProperty("sku") final String sku,
-                                          @JsonProperty("prices") List<PriceWrapper> prices,
-                                          @JsonProperty("attributes") final List<AttributeDraftWrapper> attributes,
-                                          @JsonProperty("images") List<ImageWrapper> images,
-                                          @JsonProperty("availability") final AvailabilityWrapper availability) {
-            this.sku = requireNonNull(sku);
-            this.prices = requireNonNull(prices);
-            this.attributes = requireNonNull(attributes);
-            this.images = requireNonNull(images);
-            this.availability = requireNonNull(availability);
-        }
-
-        public String getSku() {
-            return sku;
-        }
-
-        public List<PriceWrapper> getPrices() {
-            return prices;
-        }
-
-        public List<AttributeDraftWrapper> getAttributes() {
-            return attributes;
-        }
-
-        public List<ImageWrapper> getImages() {
-            return images;
-        }
-
-        public AvailabilityWrapper getAvailability() {
-            return availability;
-        }
-    }
-
-    public static class ImageWrapper {
-
-        private final String url;
-        private final ImageDimensionsWrapper dimensions;
-
-        public ImageWrapper(@JsonProperty("url") final String url,
-                            @JsonProperty("dimensions") final ImageDimensionsWrapper dimensions) {
-            this.url = requireNonNull(url);
-            this.dimensions = requireNonNull(dimensions);
-        }
-
-        public String getUrl() {
-            return url;
-        }
-
-        public ImageDimensionsWrapper getDimensions() {
-            return dimensions;
-        }
-    }
-
-    public static class ImageDimensionsWrapper {
-
-        private final Integer width;
-        private final Integer height;
-
-        public ImageDimensionsWrapper(@JsonProperty("w") final Integer width, @JsonProperty("h") final Integer height) {
-            this.width = requireNonNull(width);
-            this.height = requireNonNull(height);
-        }
-
-        public Integer getWidth() {
-            return width;
-        }
-
-        public Integer getHeight() {
-            return height;
-        }
-    }
-
-    public static class AttributeDraftWrapper {
-
-        private final String name;
-        private final Object value;
-
-        public AttributeDraftWrapper(@JsonProperty("name") final String name, @JsonProperty("value") final Object value) {
-            this.name = requireNonNull(name);
-            this.value = requireNonNull(value);
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public Object getValue() {
-            return value;
-        }
-    }
-
-    public static class AvailabilityWrapper {
-
-        private final Boolean isOnStock;
-
-        public AvailabilityWrapper(@JsonProperty("typeId") final Boolean isOnStock) {
-            this.isOnStock = isOnStock;
-        }
-    }
-
-
-    public static class ProductTypeWrapper {
-
-        private final String typeId;
-        private final String id;
-
-        public ProductTypeWrapper(@JsonProperty("typeId") final String typeId, @JsonProperty("id") final String id) {
-            this.typeId = requireNonNull(typeId);
-            this.id = requireNonNull(id);
-        }
-    }
-
-    public static class PriceWrapper {
-
-        private final PriceValueWrapper value;
-
-        public PriceWrapper(@JsonProperty("value") final PriceValueWrapper value) {
-            this.value = requireNonNull(value);
-        }
-
-        public PriceValueWrapper getValue() {
-            return value;
-        }
-    }
-
-    public static class PriceValueWrapper {
-
-        private final Double amount;
-        private final String currencyCode;
-
-        public PriceValueWrapper(@JsonProperty("amount") final Double amount,
-                                 @JsonProperty("currencyCode") final String currencyCode) {
-            this.amount = requireNonNull(amount);
-            this.currencyCode = requireNonNull(currencyCode);
-        }
-
-        public Double getAmount() {
-            return amount;
-        }
-
-        public String getCurrencyCode() {
-            return currencyCode;
-        }
     }
 }
