+++ app/models/wrapper/TextInputHintWrapper.java
@@ -1,0 +1,6 @@
+package models.wrapper;
+
+public enum TextInputHintWrapper {
+    SINGLE_LINE,
+    MULTI_LINE
+}
